refactor: Move BitSetIterator and IntIterators from util to spi package
