+++ raptor/src/main/java/org/opentripplanner/raptor/rangeraptor/transit/AccessPaths.java
@@ -15,9 +15,9 @@ import org.opentripplanner.raptor.api.model.RaptorAccessEgress;
 import org.opentripplanner.raptor.api.model.RaptorStartOnBoardAccess;
 import org.opentripplanner.raptor.api.request.RaptorProfile;
 import org.opentripplanner.raptor.spi.IntIterator;
+import org.opentripplanner.raptor.spi.IntIterators;
 import org.opentripplanner.raptor.spi.RaptorConstants;
 import org.opentripplanner.raptor.spi.SearchDirection;
-import org.opentripplanner.raptor.util.IntIterators;
 
 /**
  * This class is responsible for performing Raptor-specific functionality on access-paths. It
